# fixes #20232；indices now use `copyMem` and `moveMem`; much faster slicing and slicing assignments