--- a/lib/system/indices.nim
+++ b/lib/system/indices.nim
@@ -76,6 +76,15 @@ template spliceImpl(s, a, L, b: typed): untyped =
   # fill the hole:
   for i in 0 ..< b.len: s[a+i] = b[i]
 
+template raiseIndexErrorImpl(b: int, length: int, body: untyped) =
+  when compileOption("boundChecks"):
+    if b <= length:
+      body
+    else:
+      raiseIndexError3(b, 0, length-1)
+  else:
+    body
+
 proc `[]`*[T, U: Ordinal](s: string, x: HSlice[T, U]): string {.inline, systemRaisesDefect.} =
   ## Slice operation for strings.
   ## Returns the inclusive range `[s[x.a], s[x.b]]`:
@@ -93,10 +102,8 @@ proc `[]`*[T, U: Ordinal](s: string, x: HSlice[T, U]): string {.inline, systemRa
   else:
     when notJSnotNims:
       if L > 0:
-        if b <= s.len:
+        raiseIndexErrorImpl(b, s.len):
           copyMem(addr result[0], addr s[a], L)
-        else:
-          raiseIndexError3(b, 0, s.len-1)
     else:
       impl()
 
@@ -122,12 +129,10 @@ proc `[]=`*[T, U: Ordinal](s: var string, x: HSlice[T, U], b: string) {.systemRa
     else:
       when notJSnotNims:
         if L > 0:
-          if xb <= s.len:
+          raiseIndexErrorImpl(xb, s.len):
             when defined(nimSeqsV2):
               prepareMutation(s)
             copyMem(addr s[a], addr b[0], L)
-          else:
-            raiseIndexError3(xb, 0, s.len-1)
       else:
         impl()
   else:
@@ -151,10 +156,8 @@ proc `[]`*[Idx, T; U, V: Ordinal](a: array[Idx, T], x: HSlice[U, V]): seq[T] {.s
   else:
     when notJSnotNims and supportsCopyMem(T):
       if L > 0:
-        if xb <= a.len:
+        raiseIndexErrorImpl(xb, a.len):
           copyMem(addr result[0], addr a[Idx(xa)], sizeof(T) * L)
-        else:
-          raiseIndexError3(xb, 0, a.len-1)
     else:
       impl()
 
@@ -176,10 +179,8 @@ proc `[]=`*[Idx, T; U, V: Ordinal](a: var array[Idx, T], x: HSlice[U, V], b: ope
     else:
       when notJSnotNims and supportsCopyMem(T):
         if L > 0:
-          if xb <= a.len:
+          raiseIndexErrorImpl(xb, a.len):
             moveMem(addr a[Idx(xa)], addr b[0], sizeof(T) * L)
-          else:
-            raiseIndexError3(xb, 0, a.len-1)
       else:
         impl()
   else:
@@ -204,10 +205,8 @@ proc `[]`*[T; U, V: Ordinal](s: openArray[T], x: HSlice[U, V]): seq[T] {.systemR
   else:
     when notJSnotNims and supportsCopyMem(T):
       if L > 0:
-        if xb <= s.len:
+        raiseIndexErrorImpl(xb, s.len):
           copyMem(addr result[0], addr s[a], sizeof(T) * L)
-        else:
-          raiseIndexError3(xb, 0, s.len-1)
     else:
       impl()
 
@@ -231,10 +230,8 @@ proc `[]=`*[T; U, V: Ordinal](s: var seq[T], x: HSlice[U, V], b: openArray[T]) {
     else:
       when notJSnotNims and supportsCopyMem(T):
         if L > 0:
-          if xb <= s.len:
+          raiseIndexErrorImpl(xb, s.len):
             moveMem(addr s[a], addr b[0], sizeof(T) * L)
-          else:
-            raiseIndexError3(xb, 0, s.len-1)
       else:
         impl()
   else: