@ringabout 2023-04-14
> Just curious - wouldn't it be possible to disable implicit range checks via pushing a pragma to achieve the same result here without resorting to copyMem? Since the original issue claims that the performance hit is because of range checks

I tested it with `-d:danger`, it's still slower than `copyMem`.