@arnetheduck 2023-04-14
> I expect clang to optimize for loops into copyMem on its own these days

this happens less than you would think, specially with dynamic lengths, padding, integer wraparound rules etc